backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.10
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
flake8==7.3.0
idna==3.11
isort==7.0.0
mccabe==0.7.0
oauthlib==3.3.1
pillow==12.1.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
drf-extra-fields==3.7.0
django-filter==25.2
